Split content matching.
